+++ scripts/prepare_sector_network.py
@@ -2538,6 +2538,104 @@ def _add_smr_capacities(
         ).fillna(0.0)
 
 
+def _add_h2_storage_capacities(
+    n: pypsa.Network,
+    h2_storage_capacities: pd.DataFrame,
+) -> None:
+    """
+    Add existing H2 storage energy and charge/discharge capacities as well as optional expansion constraints.
+
+    Parameters
+    ----------
+    n : pypsa.Network#
+        The PyPSA network container object.
+    h2_storage_capacities : pd.DataFrame
+        Existing H2 storage energy and charge/discharge capacities and expansion constraints.
+
+    Returns
+    -------
+    None
+        Modifies the network object in-place by adding the H2 storage capacities.
+    """
+    logger.info("Adding H2 storage capacities.")
+
+    # H2 Store components
+    h2_stores = n.stores.carrier.isin(["H2 cavern-storage", "H2 tank-storage"])
+    h2_stores_i = n.stores[h2_stores].index
+    h2_stores_extendable_i = n.stores[h2_stores & n.stores.e_nom_extendable].index
+    # H2 charge/discharge Link components
+    h2_chargers = n.links.carrier.isin(
+        [
+            "H2 cavern-storage charger",
+            "H2 cavern-storage discharger",
+            "H2 tank-storage charger",
+            "H2 tank-storage discharger",
+        ]
+    )
+    h2_chargers_i = n.links[h2_chargers].index
+    h2_chargers_extendable_i = n.links[h2_chargers & n.links.p_nom_extendable].index
+
+    # Add capacities for H2 Stores and charge/discharge Links
+    store_caps = h2_storage_capacities.set_index("bus").e_nom
+    link_caps = pd.concat(
+        [
+            h2_storage_capacities.set_index(
+                h2_storage_capacities.bus + " charger"
+            ).p_nom_charge,
+            h2_storage_capacities.set_index(
+                h2_storage_capacities.bus + " discharger"
+            ).p_nom_discharge,
+        ]
+    )
+    n.stores.loc[h2_stores_i, ["e_nom", "e_nom_min"]] = store_caps.reindex(
+        n.stores.loc[h2_stores_i, :].index
+    ).fillna(0.0)
+    n.links.loc[h2_chargers_i, ["p_nom", "p_nom_min"]] = link_caps.reindex(
+        n.links.loc[h2_chargers_i, :].index
+    ).fillna(0.0)
+
+    # Add expansion constraints to extendable assets
+    store_caps_max = h2_storage_capacities.set_index("bus").e_nom_max
+    link_caps_max = pd.concat(
+        [
+            h2_storage_capacities.set_index(
+                h2_storage_capacities.bus + " charger"
+            ).p_nom_max_charge,
+            h2_storage_capacities.set_index(
+                h2_storage_capacities.bus + " discharger"
+            ).p_nom_max_discharge,
+        ]
+    )
+    n.stores.loc[h2_stores_extendable_i, ["e_nom_max"]] = store_caps_max.reindex(
+        n.stores.loc[h2_stores_extendable_i, :].index
+    ).fillna(0.0)
+    n.links.loc[h2_chargers_extendable_i, ["p_nom_max"]] = link_caps_max.reindex(
+        n.links.loc[h2_chargers_extendable_i, :].index
+    ).fillna(0.0)
+
+    remove_zero_capacity_non_extendable(
+        n,
+        carriers=[
+            "H2 cavern-storage",
+            "H2 tank-storage",
+            "H2 cavern-storage charger",
+            "H2 tank-storage charger",
+            "H2 cavern-storage discharger",
+            "H2 tank-storage discharger",
+        ],
+        component_types={"Store", "Link"},
+    )
+    # Drop Storage buses that do not have a store connected to it anymore
+    remaining_stores = n.stores[
+        n.stores.carrier.isin(["H2 cavern-storage", "H2 tank-storage"])
+    ].bus.unique()
+    idx = n.buses.loc[
+        n.buses.carrier.isin(["H2 cavern-storage", "H2 tank-storage"])
+        & ~n.buses.index.isin(remaining_stores)
+    ].index
+    n.remove("Bus", idx)
+
+
 def _add_other_res_profiles(
     carrier: str,
     asset_i: pd.Index,
@@ -2664,6 +2762,7 @@ def add_existing_tyndp_capacities(
     pemmdb_capacities: pd.DataFrame,
     pemmdb_profiles: pd.DataFrame,
     smr_capacities: pd.DataFrame,
+    h2_storage_capacities: pd.DataFrame,
     trajectories: pd.DataFrame,
     tyndp_renewable_carriers: list[str],
     tyndp_conventional_thermals: list[str],
@@ -2700,6 +2799,8 @@ def add_existing_tyndp_capacities(
         DataFrame containing all PEMMDB must-run and availability profiles.
     smr_capacities : pd.DataFrame
         DataFrame containing existing SMR capacities.
+    h2_storage_capacities : pd.DataFrame
+        DataFrame containing existing H2 storage capacities.
     trajectories : pd.DataFrame
         DataFrame containing the trajectories for the current pyear to attach (p_nom_min and p_nom_max).
     tyndp_renewable_carriers : list[str]
@@ -2802,10 +2903,14 @@ def add_existing_tyndp_capacities(
             )
 
     if h2_topology_tyndp:
-        logger.info("Adding SMR and SMR CC capacities.")
+        logger.info("Adding SMR, SMR CC and H2 storage capacities.")
         _add_smr_capacities(
             n=n,
             smr_capacities=smr_capacities,
+        )
+        _add_h2_storage_capacities(
+            n=n,
+            h2_storage_capacities=h2_storage_capacities,
         )
 
 
@@ -3496,9 +3601,90 @@ def add_h2_grid_tyndp(n, nodes, h2_pipes_file, interzonal_file, costs, options):
     )
 
 
+def _add_h2_stores_and_links_tyndp(
+    n: pypsa.Network,
+    storage_tech: str,
+    buses: pd.Index,
+    costs: pd.DataFrame,
+    extendable: bool,
+) -> None:
+    """
+    Adds storage Bus, Store and Charge/Discharge Link components for a given TYNDP H2 storage technology.
+
+    Parameters
+    ----------
+    n : pypsa.Network
+        The PyPSA network container object
+    storage_tech: str
+        Storage technology to add. Can be either 'cavern-storage' or 'tank-storage'
+    buses : pd.Index
+        nodes of H2 buses to add the storages to
+    costs : pd.DataFrame
+        Technology cost assumptions
+    extendable : bool
+        Whether the added storage components shall be extendable in the optimization or not.
+
+    Returns
+    -------
+    None
+        The function modifies the network object in-place by adding components.
+    """
+    bus_names = buses + f" {storage_tech}"
+
+    n.add(
+        "Bus",
+        bus_names,
+        location=buses,
+        carrier=f"H2 {storage_tech}",
+        x=n.buses.loc[list(buses)].x.values,
+        y=n.buses.loc[list(buses)].y.values,
+    )
+
+    n.add(
+        "Store",
+        bus_names,
+        bus=bus_names,
+        e_nom_extendable=extendable,
+        e_cyclic=True,
+        carrier=f"H2 {storage_tech}",
+        capital_cost=costs.at[storage_tech, "capital_cost"],
+        lifetime=costs.at[storage_tech, "lifetime"],
+    )
+
+    n.add(
+        "Link",
+        bus_names,
+        suffix=" charger",
+        bus0=buses,
+        bus1=bus_names,
+        carrier=f"H2 {storage_tech} charger",
+        efficiency=costs.at[f"{storage_tech}-charger", "efficiency"],
+        marginal_cost=costs.at[storage_tech, "marginal_cost"],
+        p_nom_extendable=extendable,
+        lifetime=costs.at[storage_tech, "lifetime"],
+    )
+
+    n.add(
+        "Link",
+        bus_names,
+        suffix=" discharger",
+        bus0=bus_names,
+        bus1=buses,
+        carrier=f"H2 {storage_tech} discharger",
+        efficiency=costs.at[f"{storage_tech}-discharger", "efficiency"],
+        marginal_cost=costs.at[storage_tech, "marginal_cost"],
+        p_nom_extendable=extendable,
+        lifetime=costs.at[storage_tech, "lifetime"],
+    )
+
+
 def add_h2_storage_tyndp(
-    n, cavern_types, h2_cavern_file, buses_h2_z1, costs, options={}
-):
+    n: pypsa.Network,
+    buses_h2_z1: pd.Index,
+    buses_h2_z2: pd.Index,
+    costs: pd.DataFrame,
+    options: dict = {},
+) -> None:
     """
     Adds TYNDP Z1 H2 tank storages and Z2 H2 cavern storages with default assumptions.
 
@@ -3506,17 +3692,15 @@ def add_h2_storage_tyndp(
     ----------
     n : pypsa.Network
         The PyPSA network container object
-    cavern_types : list
-        List of underground storage types to consider
-    h2_caverns_file : str
-        Path to CSV containing hydrogen cavern storage potentials
-    buses_h2_z1 : SimpleNamespace
-        Namespace object with spatial nodes of H2 Z1 buses
+    buses_h2_z1 : pd.Index
+        Nnodes of H2 Z1 buses
+    buses_h2_z2 : pd.Index
+        Nodes of H2 Z2 buses
     costs : pd.DataFrame
         Technology cost assumptions
     options : dict, optional
-       Dictionary of configuration options. Defaults to empty dict if not provided.
-       - hydrogen_underground_storage : bool
+        Dictionary of configuration options. Defaults to empty dict if not provided.
+       - h2_zones_tyndp : bool
 
     Returns
     -------
@@ -3524,65 +3708,26 @@ def add_h2_storage_tyndp(
         The function modifies the network object in-place by adding components.
     """
 
-    h2_caverns = pd.read_csv(h2_cavern_file, index_col=0)
-
-    # add underground hydrogen cavern storage to all H2 Z2 nodes
-    if (
-        not h2_caverns.empty
-        and options["hydrogen_underground_storage"]
-        and set(cavern_types).intersection(h2_caverns.columns)
-    ):
-        h2_caverns = h2_caverns[cavern_types].sum(axis=1)
-
-        # only use sites with at least 2 TWh potential
-        h2_caverns = h2_caverns[h2_caverns > 2]
-
-        # convert TWh to MWh
-        h2_caverns = h2_caverns * 1e6
-
-        # clip at 1000 TWh for one location
-        h2_caverns.clip(upper=1e9, inplace=True)
-
-        # group on country level
-        h2_caverns = (
-            h2_caverns.to_frame()
-            .assign(country=h2_caverns.index.map(n.buses.country).values)
-            .groupby("country")
-            .sum()
-            .loc[:, 0]
-        )
-
-        logger.info("Adding TYNDP H2 underground storage")
-
-        suffix = "H2 Z2" if options["h2_zones_tyndp"] else "H2"
-
-        n.add(
-            "Store",
-            h2_caverns.index + f" {suffix} Cavern Store",
-            bus=h2_caverns.index + f" {suffix}",
-            e_nom_extendable=True,
-            e_nom_max=h2_caverns.values,
-            e_cyclic=True,
-            carrier="H2 Store",
-            capital_cost=costs.at["hydrogen storage underground", "capital_cost"],
-            lifetime=costs.at["hydrogen storage underground", "lifetime"],
-        )
+    # Add underground hydrogen cavern storage to all H2 Z2 nodes
+    suffix = "H2 Z2" if options["h2_zones_tyndp"] else "H2"
+    logger.info(f"Adding TYNDP H2 underground storage to {suffix} nodes.")
+    _add_h2_stores_and_links_tyndp(
+        n=n,
+        storage_tech="cavern-storage",
+        buses=buses_h2_z2,
+        costs=costs,
+        extendable=options["h2_zones_tyndp"],
+    )
 
     # add overground hydrogen tank storage to all H2 Z1 nodes
     if not buses_h2_z1.empty:
-        tech = "hydrogen storage tank type 1 including compressor"
-
-        logger.info("Adding TYNDP H2 tank storage for H2 Z1")
-
-        n.add(
-            "Store",
-            buses_h2_z1 + " Tank Store",
-            bus=buses_h2_z1,
-            e_nom_extendable=True,
-            e_cyclic=True,
-            carrier="H2 Store",
-            capital_cost=costs.at[tech, "capital_cost"],
-            lifetime=costs.at[tech, "lifetime"],
+        logger.info("Adding TYNDP H2 tank storage to H2 Z1 nodes.")
+        _add_h2_stores_and_links_tyndp(
+            n=n,
+            storage_tech="tank-storage",
+            buses=buses_h2_z1,
+            costs=costs,
+            extendable=False,
         )
 
 
@@ -3590,10 +3735,8 @@ def add_h2_topology_tyndp(
     n,
     pop_layout,
     spatial,
-    h2_cavern_file,
     h2_pipes_file,
     interzonal_file,
-    cavern_types,
     costs,
     options,
     h2_demand_file,
@@ -3618,14 +3761,10 @@ def add_h2_topology_tyndp(
         Population layout with index of locations/nodes
     spatial : object
         Namespace object with spatial nodes for different carriers such as `h2_tyndp`
-    h2_cavern_file : str
-        Path to CSV file containing hydrogen cavern storage potentials
     h2_pipes_file : str
         Path to CSV file containing prepped H2 reference grid data
     interzonal_file : str
         Path to CSV file containing prepped H2 interzonal connection data
-    cavern_types : list
-        List of underground storage types to consider
     costs : pd.DataFrame
         Technology cost assumptions
     options : dict, optional
@@ -3703,9 +3842,8 @@ def add_h2_topology_tyndp(
     # add H2 storage (Z1: H2 tanks; Z2/NT H2 nodes: Salt caverns)
     add_h2_storage_tyndp(
         n=n,
-        cavern_types=cavern_types,
-        h2_cavern_file=h2_cavern_file,
         buses_h2_z1=buses_h2_z1,
+        buses_h2_z2=buses_h2_z2,
         costs=costs,
         options=options,
     )
@@ -4297,10 +4435,8 @@ def add_h2_gas_infrastructure(
             n=n,
             pop_layout=pop_layout,
             spatial=spatial,
-            h2_cavern_file=h2_cavern_file,
             h2_pipes_file=h2_pipes_file,
             interzonal_file=interzonal_file,
-            cavern_types=cavern_types,
             costs=costs,
             options=options,
             h2_demand_file=h2_demand_file,
@@ -9250,6 +9386,7 @@ if __name__ == "__main__":
     tyndp_trajectories = None
     tyndp_nuclear_profiles = None
     smr_capacities = None
+    h2_storage_capacities = None
 
     # Read in PEMMDB data, trajectories and availability profiles
     enable_pemmdb_caps = snakemake.params.electricity["pemmdb_capacities"]["enable"]
@@ -9345,6 +9482,7 @@ if __name__ == "__main__":
 
     if options["h2_topology_tyndp"]:
         smr_capacities = pd.read_csv(snakemake.input.tyndp_smr, index_col=0)
+        h2_storage_capacities = pd.read_csv(snakemake.input.tyndp_h2_storages)
 
     if enable_pemmdb_caps or options["h2_topology_tyndp"]:
         add_existing_tyndp_capacities(
@@ -9352,6 +9490,7 @@ if __name__ == "__main__":
             pemmdb_capacities=pemmdb_capacities,
             pemmdb_profiles=pemmdb_profiles,
             smr_capacities=smr_capacities,
+            h2_storage_capacities=h2_storage_capacities,
             trajectories=tyndp_trajectories,
             tyndp_renewable_carriers=tyndp_renewable_carriers,
             tyndp_conventional_thermals=tyndp_conventional_thermals,